feat: Criando Testes unitários e README

+++ src/main/java/com.geovane.transacaofinanceira/domain/model/Transacao.java
@@ -3,6 +3,8 @@ package com.geovane.transacaofinanceira.domain.model;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 public class Transacao {
@@ -10,5 +12,5 @@ public class Transacao {
     private final long id;
     private final long contaOrigem;
     private final long contaDestino;
-    private final long valor;
+    private final BigDecimal valor;
 }
